error wrapping that maintains types
This requires all ErrorCode to implement a wrap function.
ErrorCodes are normally built on top of reused base types.
Such as CodedError.

This change is useful with UserCode.
UserCode is useful as a return type of a function
so that a caller know there is a user message.
This allows the UserCdoe to be easily wrapped with additional
information while still maintaining the UserCode type.