error code wrapping with some generic reuse
The goal is to wrap an ErrorCode
while maintaining its type.
This allows for using UserCode as a marke type
that a user message will be returned
and still allows for wrapping the error
with extra information.

In-place modification accomplishes this,
and will be used if available (via ErrorWrapper)
However, requiring this in the ErrorCode interface
is too burdensome.
Instead fallback to a wrapper type
that will be different per-interface.
There are different wrapping functions for each interface
to maintain the type and not downcast to ErrorCode.